[Refactor] Move files to subdirectory
