+++ src/stories/organisms/Page.stories.ts
@@ -4,7 +4,7 @@ import { within, userEvent, expect } from '@storybook/test';
 import { Page } from './Page';
 
 const meta = {
-  title: 'Example/Page',
+  title: 'Organisms/Page',
   component: Page,
   parameters: {
     // More on how to position stories at: https://storybook.js.org/docs/configure/story-layout
